requirements.txt
Django==4.1.1
django-extensions==3.2.1
dj-rest-auth==2.2.5
djangorestframework==3.13.1
flake8==5.0.4
django-allauth==0.51.0
